Move chapter 2 listings to standard file layout.
