--- src/main/kotlin/chapter2/Listing_2_4.kt
@@ -1,16 +1,0 @@
-package chapter2
-
-object Listing_2_4 {
-
-    //tag::init[]
-    fun findFirst(ss: Array<String>, key: String): Int {
-        tailrec fun loop(n: Int): Int =
-            when {
-                n >= ss.size -> -1 // <1>
-                ss[n] == key -> n // <2>
-                else -> loop(n + 1) // <3>
-            }
-        return loop(0) // <4>
-    }
-    //end::init[]
-}
